Tweak AuthProviderInner
