+++ frontend/src/context/AuthProviderInner.tsx
@@ -24,55 +24,68 @@ export const AuthProviderInner = ({ children }: { children: ReactNode }) => {
     isMounted.current = true;
 
     const fetchAdminStatus = async (accessToken: string) => {
-      const res = await fetch('/api/users/me', {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      });
-
-      const body = await res.json().catch(() => null);
-
-      if (!isMounted.current) return;
-      setIsAdmin(body?.isAdmin === true);
-      setIsSuperuser(body?.isSuperuser === true);
-
-      setUser((prev) =>
-        prev
-          ? {
-              ...prev,
-              role: body?.role ?? null,
-            }
-          : null
-      );
-    };
-
-    const updateAuthState = async () => {
-      const {
-        data: { session },
-      } = await supabase.auth.getSession();
-
-      if (!isMounted.current) return;
-
-      if (session?.user) {
-        setUser({
-          name: session.user.user_metadata.full_name ?? session.user.email ?? 'User',
-          email: session.user.email ?? '',
-          role: null,
+      try {
+        const res = await fetch('/api/users/me', {
+          headers: {
+            Authorization: `Bearer ${accessToken}`,
+          },
         });
 
-        if (session.access_token) {
-          await fetchAdminStatus(session.access_token);
-        } else {
-          setIsAdmin(false);
-          setIsSuperuser(false);
-        }
-      } else {
-        setUser(null);
+        const body = await res.json().catch(() => null);
+
+        if (!isMounted.current) return;
+        setIsAdmin(body?.isAdmin === true);
+        setIsSuperuser(body?.isSuperuser === true);
+
+        setUser((prev) =>
+          prev
+            ? {
+                ...prev,
+                role: body?.role ?? null,
+              }
+            : null
+        );
+      } catch {
+        if (!isMounted.current) return;
         setIsAdmin(false);
         setIsSuperuser(false);
       }
+    };
 
-      setIsInitialized(true);
+    const updateAuthState = async () => {
+      try {
+        const {
+          data: { session },
+        } = await supabase.auth.getSession();
+
+        if (!isMounted.current) return;
+
+        if (session?.user) {
+          setUser({
+            name: session.user.user_metadata.full_name ?? session.user.email ?? 'User',
+            email: session.user.email ?? '',
+            role: null,
+          });
+
+          if (session.access_token) {
+            await fetchAdminStatus(session.access_token);
+          } else {
+            setIsAdmin(false);
+            setIsSuperuser(false);
+          }
+        } else {
+          setUser(null);
+          setIsAdmin(false);
+          setIsSuperuser(false);
+        }
+      } catch {
+        if (!isMounted.current) return;
+        setUser(null);
+        setIsAdmin(false);
+        setIsSuperuser(false);
+      } finally {
+        if (isMounted.current) setIsInitialized(true);
+      }
     };
 
     updateAuthState();
